# Doctors/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"LLinks": {
				"properties": {
					"rel": {
						"type": "string"
					},
					"href": {
						"type": "string"
					}
				},
				"required": [
					"rel",
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShowRegion": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"nameKey": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/LLinks"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"name",
					"nameKey",
					"code",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Links3": {
				"properties": {
					"rel": {
						"type": "string"
					},
					"href": {
						"type": "string"
					}
				},
				"required": [
					"rel",
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResMetaData": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"dataSource": {
						"type": "string"
					},
					"extractDate": {
						"type": "string"
					},
					"itemCount": {
						"type": "string"
					},
					"returncode": {
						"type": "string"
					},
					"totalCount": {
						"type": "string"
					},
					"totalPages": {
						"type": "string"
					},
					"pageSize": {
						"type": "string"
					},
					"pageNumber": {
						"type": "string"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Links3"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"dataSource",
					"extractDate",
					"itemCount",
					"returncode",
					"totalCount",
					"totalPages",
					"pageSize",
					"pageNumber",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CombineArray": {
				"properties": {
					"region": {
						"items": {
							"$ref": "#/components/schemas/ShowRegion"
						},
						"type": "array"
					},
					"meta": {
						"items": {
							"$ref": "#/components/schemas/ResMetaData"
						},
						"type": "array"
					}
				},
				"required": [
					"region",
					"meta"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Links": {
				"properties": {
					"rel": {
						"type": "string"
					},
					"href": {
						"type": "string"
					}
				},
				"required": [
					"rel",
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Region": {
				"properties": {
					"name": {
						"type": "string"
					},
					"nameKey": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Links"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"nameKey",
					"code",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Links222": {
				"properties": {
					"rel": {
						"type": "string"
					},
					"href": {
						"type": "string"
					}
				},
				"required": [
					"rel",
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReqMetaDataa": {
				"properties": {
					"dataSource": {
						"type": "string"
					},
					"extractDate": {
						"type": "number",
						"format": "double"
					},
					"itemCount": {
						"type": "number",
						"format": "double"
					},
					"returncode": {
						"type": "number",
						"format": "double"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"pageNumber": {
						"type": "number",
						"format": "double"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Links222"
						},
						"type": "array"
					}
				},
				"required": [
					"dataSource",
					"extractDate",
					"itemCount",
					"returncode",
					"totalCount",
					"totalPages",
					"pageSize",
					"pageNumber",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pagination": {
				"properties": {
					"pageNumber": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"pageNumber",
					"pageSize"
				],
				"type": "object",
				"additionalProperties": false
			},
			"getRegionById": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"sortingRegions": {
				"properties": {
					"SortBy": {
						"type": "string"
					}
				},
				"required": [
					"SortBy"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShowImageee": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReqImageUpload": {
				"properties": {
					"image": {
						"type": "string"
					}
				},
				"required": [
					"image"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "admin_token",
				"in": "header"
			},
			"jwt": {
				"type": "http",
				"scheme": "Bearer",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "swaggerapis2",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/region/save-region": {
			"post": {
				"operationId": "SaveData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShowRegion"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Region"
							}
						}
					}
				}
			}
		},
		"/region/save-meta": {
			"post": {
				"operationId": "SaveeMetaData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResMetaData"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqMetaDataa"
							}
						}
					}
				}
			}
		},
		"/region/pagination": {
			"post": {
				"operationId": "Pagination",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CombineArray"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Pagination"
							}
						}
					}
				}
			}
		},
		"/region/get-region-by-id": {
			"post": {
				"operationId": "GetRegionBy",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShowRegion"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/getRegionById"
							}
						}
					}
				}
			}
		},
		"/region/sorting": {
			"post": {
				"operationId": "Sorting",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShowRegion"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/sortingRegions"
							}
						}
					}
				}
			}
		},
		"/region/upload-image": {
			"post": {
				"operationId": "ImageUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShowImageee"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqImageUpload"
							}
						}
					}
				}
			}
		},
		"/region/get-region-by-id/{id}": {
			"post": {
				"operationId": "GetRegionBy2",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShowRegion"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/region/paginations/region?page={page}&limit={limit}": {
			"get": {
				"operationId": "PaginationQuery",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CombineArray"
								}
							}
						}
					}
				},
				"tags": [
					"region"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}